[enh] Each engine can define a "search" function to parse the request the way it wants (to implement an offline engine for example) The "can_accept_search_query" function can be defined too. The default behavior remains the same as before ( see searx/engines/__base__.py )
